Extract daemon API ownership
