+++ openspec/changes/improve-macos-app-architecture-maintainability/tasks.md
@@ -23,7 +23,7 @@
 
 - [x] 4.1 Move mobile or legacy console UI from `ContentView.swift` into a clearly named console/mobile folder while preserving the non-macOS app entry path.
 - [x] 4.2 Move chat/timeline/pending-yield value types into model files or console-specific model files.
-- [ ] 4.3 Split daemon API DTOs and `AlanAPIClient` ownership from console view model and SwiftUI screens.
+- [x] 4.3 Split daemon API DTOs and `AlanAPIClient` ownership from console view model and SwiftUI screens.
 - [ ] 4.4 Extract daemon event polling/reading and event-to-UI-state reduction from `AlanConsoleViewModel` so event mapping is testable without rendering the full console UI.
 
 ## 5. Shell Model And Controller Ownership
